+++ pyproject.toml
@@ -27,6 +27,7 @@ pre-commit = "^3.7.1"
 mypy = "^1.10.0"
 types-requests = "^2.32.0.20240602"
 ipykernel = "^6.29.5"
+torch = "^2.3.1"
 
 [tool.ruff]
 line-length = 100
