+++ scripts/benchmark_swish_cuda.py
